requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.6
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
whitenoise==3.3.1
yelpapi==2.0.1
